Feature: moved addProductToCart, updated navbar link to cart
