+++ client/store/cart.js
@@ -4,6 +4,7 @@ import axios from 'axios'
  * ACTION TYPES
  */
 const GOT_CART_ITEMS = 'GOT_CART_ITEMS'
+const ADDED_PRODUCT_TO_CART = 'ADDED_PRODUCT_TO_CART'
 const DELETE_PRODUCT_FROM_CART = 'DELETE_PRODUCT_FROM_CART'
 // const INCREMENT_ITEM_QUANTITY = 'INCREMENT_ITEM_QUANTITY'
 // const DECREMENT_ITEM_QUANTITY = 'DECREMENT_ITEM_QUANTITY'
@@ -18,9 +19,15 @@ const initialState = {
 /**
  * ACTION CREATORS
  */
+
 const gotCartItems = cartItems => ({
   type: GOT_CART_ITEMS,
   cartItems
+})
+
+const addedProductToCart = productId => ({
+  type: ADDED_PRODUCT_TO_CART,
+  productId
 })
 
 const deletedProductFromCart = updatedCart => ({
@@ -46,6 +53,19 @@ export const getCartItems = userId => async dispatch => {
     const {data} = await axios.get(`/api/cart/${userId}`)
     console.log('DATA IN THUNK: ', data)
     dispatch(gotCartItems(data))
+  } catch (error) {
+    console.error(error)
+  }
+}
+
+export const addProductToCart = (userId, productId) => async dispatch => {
+  try {
+    console.log('addProdToCart')
+    const res = await axios.post('/api/cart/order', {
+      userId: userId,
+      productId: productId
+    })
+    dispatch(addedProductToCart(res.data))
   } catch (error) {
     console.error(error)
   }
@@ -89,6 +109,8 @@ export default function(state = initialState, action) {
   switch (action.type) {
     case GOT_CART_ITEMS:
       return {...state, cartItems: action.cartItems}
+    case ADDED_PRODUCT_TO_CART:
+      return {...state, currentProduct: action.productId}
     // case DELETE_PRODUCT_FROM_CART:
     //   return {
     //     ...state,
