+++ client/store/product.js
@@ -5,14 +5,13 @@ import axios from 'axios'
  */
 const GOT_PRODUCTS = 'GOT_PRODUCTS'
 const GOT_SINGLE_PRODUCT = 'GOT_SINGLE_PRODUCT'
-const ADD_PRODUCT_TO_CART = 'ADD_PRODUCT_TO_CART'
 
 /**
  * INITIAL STATE
  */
 const initialState = {
   products: [],
-  currentProduct: []
+  currentProduct: {}
 }
 
 /**
@@ -25,10 +24,6 @@ const gotProducts = products => ({
 const gotSingleProduct = currentProduct => ({
   type: GOT_SINGLE_PRODUCT,
   currentProduct
-})
-const addedProductToCart = productId => ({
-  type: ADD_PRODUCT_TO_CART,
-  productId
 })
 
 /**
@@ -52,19 +47,6 @@ export const getSingleProduct = id => async dispatch => {
   }
 }
 
-export const addProductToCart = (userId, productId) => async dispatch => {
-  try {
-    console.log('addProdToCart')
-    const res = await axios.post('/api/cart/order', {
-      userId: userId,
-      productId: productId
-    })
-    dispatch(addedProductToCart(res.data))
-  } catch (error) {
-    console.error(error)
-  }
-}
-
 /**
  * REDUCER
  */
@@ -74,8 +56,6 @@ export default function(state = initialState, action) {
       return {...state, products: action.products}
     case GOT_SINGLE_PRODUCT:
       return {...state, currentProduct: action.currentProduct}
-    case ADD_PRODUCT_TO_CART:
-      return {...state, currentProduct: action.productId}
     default:
       return state
   }
